requirements.txt
Flask
python-dotenv
langchain
langchain_community
langchain_huggingface
langchain_groq
faiss-cpu
requests